Formatting and other fixes.
I changed the uView display code slightly based on some basic understanding of display control and a skim of the docs.  If I'm wrong, fix it.  Also, the setup block seemed to be missing a curly brace.

AvE, please don't try to invent your own formatting with spaces and such.  Maybe you can read it, but it's not standard, and I can't really read it.